[compiler] Fix PrintParameter of LoadLane
The laneidx of LoadLaneParameters will be taken as char type when
PrintParameter. Fix operator<< to print laneidx with integer type.

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3878768
Auto-Submit: Yolanda Chen <yolanda.chen@intel.com>
Commit-Queue: Yolanda Chen <yolanda.chen@intel.com>
Cr-Commit-Position: refs/heads/main@{#83022}

+++ src/compiler/machine-operator.cc
@@ -156,8 +156,8 @@ size_t hash_value(LoadLaneParameters params) {
 }
 
 std::ostream& operator<<(std::ostream& os, LoadLaneParameters params) {
-  return os << "(" << params.kind << " " << params.rep << " " << params.laneidx
-            << ")";
+  return os << "(" << params.kind << " " << params.rep << " "
+            << static_cast<uint32_t>(params.laneidx) << ")";
 }
 
 LoadLaneParameters const& LoadLaneParametersOf(Operator const* op) {
